# backend/openapi.json
{
  "components" : {
    "schemas" : {
...
      "CareTeamRequest" : {
        "properties" : {
          "members" : {
            "items" : {
              "$ref" : "#/components/schemas/Member"
            },
            "type" : "array"
          },
          "name" : {
            "type" : "string"
          },
          "patientId" : {
            "type" : "string"
          }
        },
        "required" : [ "members", "name", "patientId" ],
        "type" : "object"
      },
      "CareTeamSummary" : {
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "members" : {
            "items" : {
              "$ref" : "#/components/schemas/Member"
            },
            "type" : "array"
          },
          "name" : {
            "type" : "string"
          },
          "patientId" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "GoalRequest" : {
        "properties" : {
          "addressesConditionId" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          },
          "patientId" : {
            "type" : "string"
          },
          "priority" : {
            "type" : "string"
          },
          "targetDate" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "required" : [ "description", "patientId" ],
        "type" : "object"
      },
      "GoalSummary" : {
        "properties" : {
          "achievementStatus" : {
            "type" : "string"
          },
          "addressesConditionId" : {
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "lifecycleStatus" : {
            "type" : "string"
          },
          "patientId" : {
            "type" : "string"
          },
          "priority" : {
            "type" : "string"
          },
          "targetDate" : {
            "format" : "date",
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ImmunizationRequest" : {
        "properties" : {
          "doseNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "encounterId" : {
            "type" : "string"
          },
          "lotNumber" : {
            "type" : "string"
          },
          "note" : {
            "type" : "string"
          },
          "occurrenceDate" : {
            "format" : "date",
            "type" : "string"
          },
          "patientId" : {
            "type" : "string"
          },
          "performerId" : {
            "type" : "string"
          },
          "route" : {
            "type" : "string"
          },
          "site" : {
            "type" : "string"
          },
          "vaccineCode" : {
            "type" : "string"
          },
          "vaccineDisplay" : {
            "type" : "string"
          },
          "vaccineSystem" : {
            "type" : "string"
          }
        },
        "required" : [ "occurrenceDate", "patientId", "performerId", "vaccineCode", "vaccineDisplay", "vaccineSystem" ],
        "type" : "object"
      },
      "ImmunizationSummary" : {
        "properties" : {
          "doseNumber" : {
            "format" : "int32",
            "type" : "integer"
          },
          "id" : {
            "type" : "string"
          },
          "lotNumber" : {
            "type" : "string"
          },
          "occurrenceDate" : {
            "format" : "date",
            "type" : "string"
          },
          "patientId" : {
            "type" : "string"
          },
          "site" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "statusReason" : {
            "type" : "string"
          },
          "vaccine" : {
            "type" : "string"
          },
          "vaccineCode" : {
            "type" : "string"
          },
          "vaccineSystem" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Member" : {
        "properties" : {
          "practitionerId" : {
            "type" : "string"
          },
          "roleCode" : {
            "type" : "string"
          },
          "roleDisplay" : {
            "type" : "string"
          }
        },
        "required" : [ "practitionerId", "roleCode", "roleDisplay" ],
        "type" : "object"
      },
...
    },
    "securitySchemes" : {
      "bearer-jwt" : {
        "bearerFormat" : "JWT",
        "description" : "Keycloak issued access token carrying realm roles and SMART scopes.",
        "scheme" : "bearer",
        "type" : "http"
      }
    }
  },
  "info" : {
    "description" : "Task oriented endpoints for the Zantrix clinical workspace and patient portal. Clinical state is FHIR R4 and is reached through the guarded gateway; these endpoints compose it and never form a parallel data model. The conformant FHIR surface is published separately at /fhir/R4.",
    "license" : {
      "name" : "GNU Affero General Public License v3.0",
      "url" : "https://www.gnu.org/licenses/agpl-3.0.txt"
    },
    "title" : "Zantrix Application API",
    "version" : "v1"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/v1/care/goals" : {
      "get" : {
        "operationId" : "goals",
        "parameters" : [ {
          "in" : "query",
          "name" : "patientId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "includeClosed",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/GoalSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "care-coordination-controller" ]
      },
      "post" : {
        "operationId" : "addGoal",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/GoalRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GoalSummary"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "care-coordination-controller" ]
      }
    },
    "/api/v1/care/goals/{id}/close" : {
      "post" : {
        "operationId" : "closeGoal",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "patientId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "outcome",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/GoalSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "care-coordination-controller" ]
      }
    },
    "/api/v1/care/teams" : {
      "get" : {
        "operationId" : "teams",
        "parameters" : [ {
          "in" : "query",
          "name" : "patientId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CareTeamSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "care-coordination-controller" ]
      },
      "post" : {
        "operationId" : "defineTeam",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CareTeamRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CareTeamSummary"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "care-coordination-controller" ]
      }
    },
    "/api/v1/care/teams/{id}/stand-down" : {
      "post" : {
        "operationId" : "standDownTeam",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "patientId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CareTeamSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "care-coordination-controller" ]
      }
    },
...
    "/api/v1/immunizations" : {
      "get" : {
        "operationId" : "list_6",
        "parameters" : [ {
          "in" : "query",
          "name" : "patientId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "includeEnteredInError",
          "required" : false,
          "schema" : {
            "default" : false,
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/ImmunizationSummary"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "immunization-controller" ]
      },
      "post" : {
        "operationId" : "record_1",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ImmunizationRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImmunizationSummary"
                }
              }
            },
            "description" : "Created"
          }
        },
        "tags" : [ "immunization-controller" ]
      }
    },
    "/api/v1/immunizations/{id}/entered-in-error" : {
      "post" : {
        "operationId" : "markEnteredInError",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "patientId",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "reason",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ImmunizationSummary"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "immunization-controller" ]
      }
    },
...
  },
  "security" : [ {
    "bearer-jwt" : [ ]
  } ],
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}